+++ luals-addon/factorio/library/runtime/LuaObject.lua
@@ -1,10 +1,5 @@
 ---@meta _
 
----@param object LuaObject
----@return LuaObject.object_name object_name
----@nodiscard
-function __object_name(object) end
-
----@class LuaObject: userdata
+---@class LuaObject.base: userdata
 ---@field valid boolean
 ---@field object_name LuaObject.object_name