+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   * `LuaObject` inherits from `userdata` (#170)
   * Add `feature_flags` global var for data stage
   * Omit `GuiStyle` from `AnyPrototype`, to reduce incorrect type deduction on data:extend's argument
+  * `LuaObject`s now narrow on `object_name` without plugin assistance
 * Debugger:
   * fix incorrect error when remote.call is called with invalid arguments
   * fix incorrect nesting of values in Output events